refactor: RootLayout
